refactor(Prompts): move list components to lists/ subdirectory
